Get scalacheck out of Configuration.scala.
