+++ scalatest-test/src/test/scala/org/scalatest/prop/PropertyCheckConfigurationHelperSuite.scala
@@ -20,7 +20,7 @@ import org.scalactic.anyvals._
 
 class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
 
-  import org.scalatest.prop.Configuration._
+  import org.scalatest.check.ScalaCheckConfiguration._
 
   val DefaultMinSuccessful: PosInt = 9
   val PassedMinSuccessful: PosInt = 3
